# cmd/docs/swagger.yaml
basePath: /
definitions:
  domain.AccountType:
    enum:
    - ASSET
    - LIABILITY
    - EQUITY
    - INCOME
    - EXPENSE
    type: string
    x-enum-varnames:
    - Asset
    - Liability
    - Equity
    - Income
    - Expense
  domain.TransactionType:
    enum:
    - DEBIT
    - CREDIT
    type: string
    x-enum-varnames:
    - Debit
    - Credit
...
  dto.AccountResponse:
    properties:
      accountID:
        type: string
      accountType:
        $ref: '#/definitions/domain.AccountType'
      createdAt:
        type: string
      createdBy:
        description: UserID
        type: string
      currencyCode:
        type: string
      description:
        type: string
      isActive:
        type: boolean
      lastUpdatedAt:
        type: string
      lastUpdatedBy:
        description: UserID
        type: string
      name:
        type: string
      parentAccountID:
        type: string
      workplaceID:
        type: string
    type: object
...
  dto.CreateAccountRequest:
    properties:
      accountType:
        allOf:
        - $ref: '#/definitions/domain.AccountType'
        enum:
        - ASSET
        - LIABILITY
        - EQUITY
        - INCOME
        - EXPENSE
      currencyCode:
        type: string
      description:
        type: string
      name:
        type: string
      parentAccountID:
        description: Optional, must be UUID if provided
        type: string
    required:
    - accountType
    - currencyCode
    - name
    type: object
...
  dto.ListAccountsResponse:
    properties:
      accounts:
        items:
          $ref: '#/definitions/dto.AccountResponse'
        type: array
    type: object
...
  dto.ListTransactionsResponse:
    properties:
      transactions:
        items:
          $ref: '#/definitions/dto.TransactionResponse'
        type: array
    type: object
...
  dto.TransactionResponse:
    properties:
      accountID:
        type: string
      amount:
        description: Always positive
        type: number
      createdAt:
        type: string
      createdBy:
        type: string
      currencyCode:
        type: string
      journalID:
        type: string
      notes:
        type: string
      transactionID:
        type: string
      transactionType:
        $ref: '#/definitions/domain.TransactionType'
    type: object
  dto.UpdateAccountRequest:
    properties:
      description:
        type: string
      isActive:
        type: boolean
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for MMA backend.
  title: MMA Backend API
  version: "1.0"
paths:
...
  /workplaces/{workplace_id}/accounts:
    get:
      description: Retrieves a list of accounts for the specified workplace if the
        user is a member.
      parameters:
      - description: Workplace ID
        in: path
        name: workplace_id
        required: true
        type: string
      - default: 20
        description: Limit number of results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListAccountsResponse'
        "400":
          description: Missing Workplace ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (User not part of workplace)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to list accounts
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List accounts for current user in workplace
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Creates a new account within the specified workplace.
      parameters:
      - description: Workplace ID
        in: path
        name: workplace_id
        required: true
        type: string
      - description: Account details
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AccountResponse'
        "400":
          description: Invalid input or missing Workplace ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (User cannot create in this workplace)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create account
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create account in workplace
      tags:
      - accounts
  /workplaces/{workplace_id}/accounts/{id}:
    delete:
      description: Marks an account as inactive within a specified workplace.
      parameters:
      - description: Workplace ID
        in: path
        name: workplace_id
        required: true
        type: string
      - description: Account ID to deactivate
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Missing Workplace or Account ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (User cannot deactivate)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found in this workplace
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict (e.g., already inactive)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to deactivate account
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate account in workplace
      tags:
      - accounts
    get:
      description: Retrieves details for a specific account by its ID within a workplace.
      parameters:
      - description: Workplace ID
        in: path
        name: workplace_id
        required: true
        type: string
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AccountResponse'
        "400":
          description: Missing Workplace or Account ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (User not part of workplace)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found in this workplace
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to retrieve account
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get account by ID from workplace
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: Updates details for a specific account within a workplace.
      parameters:
      - description: Workplace ID
        in: path
        name: workplace_id
        required: true
        type: string
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account details to update
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AccountResponse'
        "400":
          description: Invalid input or missing IDs
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (User cannot update)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found in this workplace
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update account
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update account in workplace
      tags:
      - accounts
  /workplaces/{workplace_id}/accounts/{id}/transactions:
    get:
      description: Retrieves a paginated list of transactions associated with a specific
        account within a workplace.
      parameters:
      - description: Workplace ID
        in: path
        name: workplace_id
        required: true
        type: string
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - default: 20
        description: Limit number of results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListTransactionsResponse'
        "400":
          description: Missing Workplace/Account ID or invalid query params
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (User not part of workplace)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found in this workplace
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to list transactions
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List transactions for an account in a workplace
      tags:
      - accounts
...
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"